requirements.txt
llama-index>=0.8.54
langchain>=0.0.310
openai>=1.6.1
faiss-cpu>=1.7.4
numpy>=1.23.5
pandas>=1.5.3
python-dotenv>=0.19.0
transformers>=4.31.0
langchain-experimental>=0.0.29
